requirements.txt
streamlit==1.31.0
snowflake-snowpark-python==1.12.1
pandas==2.2.2